+++ LostTimeDB/LostTimeDbUp/Scripts/Process/Quest/170703111121_CreateProcessDeleteQuestByQuestID.sql
@@ -1,0 +1,10 @@
+/* Migration Script */
+
+CREATE PROCEDURE DeleteQuestByQuestID
+(
+	@QuestID int
+)
+AS
+BEGIN
+	DELETE FROM Quest WHERE QuestID = @QuestID;
+END