+++ LostTimeDB/LostTimeDbUp/Scripts/Tables/Quest/170703110127_CreateTableQuest.sql
@@ -1,0 +1,9 @@
+/* Migration Script */
+create table Quest
+(
+	QuestID INT IDENTITY(1, 1) PRIMARY KEY,
+	QuestTitle NVARCHAR(64) NOT NULL,
+	QuestData nvarchar(max) not null,
+	QuestLastEdit datetime2 not null,
+	AuthorID int not null,
+);