+++ LostTimeDB/LostTimeDbUp/Scripts/Views/Quest/170703112103_CreateViewTableQuestGetAllQuest.sql
@@ -1,0 +1,13 @@
+/* Migration Script */
+
+create view GetAllQuest
+AS
+	SELECT
+		QuestID = s.QuestID,
+		QuestTitle = s.QuestTitle,
+		QuestData = s.QuestData,
+		QuestLastEdit = s.QuestLastEdit,
+		AuthorID = s.AuthorID
+	FROM Quest s
+	where s.QuestID <> 0;
+